+++ src/main/java/roomescape/controller/page/UserPageController.java
@@ -1,11 +1,11 @@
-package roomescape.controller;
+package roomescape.controller.page;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
 
 @Controller
 public class UserPageController {
-    
+
     @GetMapping("/reservation")
     public String userReservationPage() {
         return "reservation";
