+++ src/main/java/roomescape/controller/timeslot/dto/TimeSlotResponse.java
@@ -5,20 +5,20 @@ import java.util.List;
 import roomescape.model.TimeSlot;
 
 public record TimeSlotResponse(
-    Long id,
-    LocalTime startAt
+        Long id,
+        LocalTime startAt
 ) {
 
     public static TimeSlotResponse from(TimeSlot timeSlot) {
         return new TimeSlotResponse(
-            timeSlot.id(),
-            timeSlot.startAt()
+                timeSlot.id(),
+                timeSlot.startAt()
         );
     }
 
     public static List<TimeSlotResponse> from(List<TimeSlot> timeSlots) {
         return timeSlots.stream()
-            .map(TimeSlotResponse::from)
-            .toList();
+                .map(TimeSlotResponse::from)
+                .toList();
     }
 }
