+++ src/main/java/roomescape/repository/ThemeRepository.java
@@ -1,6 +1,5 @@
 package roomescape.repository;
 
-import java.time.LocalDate;
 import java.util.List;
 import java.util.Optional;
 import roomescape.model.Theme;
