+++ src/main/java/roomescape/service/dto/AvailableTimeSlotDto.java
@@ -4,16 +4,16 @@ import java.time.LocalTime;
 import roomescape.model.TimeSlot;
 
 public record AvailableTimeSlotDto(
-    Long id,
-    LocalTime startAt,
-    Boolean alreadyBooked
+        Long id,
+        LocalTime startAt,
+        Boolean alreadyBooked
 ) {
 
     public static AvailableTimeSlotDto from(TimeSlot timeSlot, Boolean alreadyBooked) {
         return new AvailableTimeSlotDto(
-            timeSlot.id(),
-            timeSlot.startAt(),
-            alreadyBooked
+                timeSlot.id(),
+                timeSlot.startAt(),
+                alreadyBooked
         );
     }
 }
